requirements.txt
asgiref == 3.4.1
Django == 3.2.6
django-cors-headers == 3.8.0
django-rest-framework == 0.1.0
django-simple-history == 3.0.0
djangorestframework == 3.12.4
drf-renderer-xlsx == 0.4.1
et-xmlfile == 1.1.0
openpyxl == 3.0.7
pip == 21.1.3
pytz == 2021.1
setuptools == 56.0.0
sqlparse == 0.4.1